refactor: BackendStack을 apps 폴더 아래로 변경
